Move Attributes to its own file
