+++ ktoml-core/src/commonMain/kotlin/com/akuleshov7/ktoml/encoders/TomlArrayEncoder.kt
@@ -20,7 +20,7 @@ public class TomlArrayEncoder internal constructor(
     private val rootNode: TomlNode,
     private val parent: TomlAbstractEncoder?,
     elementIndex: Int,
-    attributes: Attributes,
+    attributes: TomlEncoderAttributes,
     inputConfig: TomlInputConfig,
     outputConfig: TomlOutputConfig,
     serializersModule: SerializersModule
@@ -44,7 +44,7 @@ public class TomlArrayEncoder internal constructor(
     public constructor(
         rootNode: TomlNode,
         elementIndex: Int,
-        attributes: Attributes,
+        attributes: TomlEncoderAttributes,
         inputConfig: TomlInputConfig,
         outputConfig: TomlOutputConfig,
         serializersModule: SerializersModule
